Update p0_user_stories.md
Upvoting a couple stories.

_"I want 2 apps in different namespaces to connect via namespace NAME"_
I remember feeling surprised that this one wasn't possible out of the box and it's a pain in situations where you can't infer or update the labels on the namespace.

_"I dont want to look at 10 or 100 policies to figure out whether I have the right allow rules"_
This would be nice. :) Especially if you're working on an unfamiliar, shared cluster and aren't familiar with all the policies they've applied.